+++ docs/changelog.rst
@@ -27,6 +27,8 @@ In progress (unreleased)
   to :attr:`mitogen.core.ImportPolicy.unsuitables`
 * :gh:issue:`1451` :mod:`mitogen`: Always delegate modules in
   :attr:`mitogen.core.ImportPolicy.unsuitables` to Python's own importers
+* :gh:issue:`1451` :mod:`mitogen`: Add discovered stdlib module names to
+  :attr:`mitogen.core.ImportPolicy.unsuitables`
 
 
 v0.3.41 (2026-02-10)